1- Setup a compose Navigation 2- create list and details screens with sample ui

+++ composeApp/src/commonMain/kotlin/data/mappers/coffeeItemMapper.kt
@@ -4,5 +4,5 @@ import data.dto.CoffeeItemDTO
 import domain.CoffeeItem
 
 fun CoffeeItemDTO.toCoffeeItem(): CoffeeItem {
-    return CoffeeItem(id, name, description, price, image_url)
+    return CoffeeItem(id, name, description, price.toFloat(), image_url)
 }
